src/auth/utils.py
from passlib.context import CryptContext
from datetime import timedelta, datetime
from src.config import Config
import jwt
import uuid, logging

password_contxt=CryptContext(
    schemes=['bcrypt']
)

ACCESS_TOKEN_EXPIRY = 3600

def generate_password_hash(password:str) -> str:
    hash = password_contxt.hash(password)
    return hash

def verify_hash(password:str, hash:str) -> bool:
    return password_contxt.verify(password,hash)

def verify_password(password:str,password_hash:str) -> bool:
    return password_contxt.verify(password, password_hash)

def create_access_token(user_data:dict,expiry:timedelta,refresh:bool=False):
    payload = {}
    payload['user'] = user_data
    payload['exp'] = datetime.now() + expiry if expiry is not None else timedelta(seconds=ACCESS_TOKEN_EXPIRY)
    payload['jti'] = str(uuid.uuid4())
    payload['refresh'] = refresh
    token = jwt.encode(
        payload=payload, key=Config.JWT_SECRET, algorithm=Config.JWT_ALGORITHM
    )

    token=jwt.encode(
        payload=payload,
        key=Config.JWT_SECRET,
        algorithm=Config.JWT_ALGORITHM
    )   
    return token 

def decode_token(token:str):
    try:
        token_data =jwt.decode(
            jwt=token,
            key=Config.JWT_SECRET,
            algorithms=Config.JWT_ALGORITHM,
        )
        return token_data
    
    except jwt.PyJWTError as e:
        logging.exception(e)
        return None