requirements.txt
# Core Framework
fastapi==0.118.3
uvicorn==0.37.0
python-multipart==0.0.20

# Database
SQLAlchemy==2.0.45
sqlmodel==0.0.27
asyncpg==0.30.0
alembic==1.17.2

# Data Validation
pydantic==2.12.0
pydantic-settings==2.12.0
email-validator==2.3.0

# Authentication & Security
PyJWT==2.8.1
passlib==1.7.4
bcrypt==4.1.3

# Redis
aioredis==2.0.1

# Environment Variables
python-dotenv==1.2.1

# Async Support
anyio==4.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
sniffio==1.3.1
websockets==15.0.1
wheel==0.45.1
passlib==1.7.4
